+++ src/demo/luigi.c
@@ -157,7 +157,6 @@ int luigi_demo(struct notcurses* nc){
   struct ncvisual_options vopts = {
     .n = notcurses_stddim_yx(nc, &rows, &cols),
     .scaling = NCSCALE_STRETCH,
-    .blitter = NCBLIT_2x2,
   };
   if(ncvisual_render(nc, nv, &vopts) == NULL){
     return -1;