+++ src/poc/visual.cpp
@@ -20,7 +20,7 @@ int main(int argc, char** argv){
     file = argv[1];
   }
   notcurses_options opts{};
-  opts.loglevel = NCLOGLEVEL_TRACE;
+  //opts.loglevel = NCLOGLEVEL_TRACE;
   opts.flags = NCOPTION_INHIBIT_SETLOCALE | NCOPTION_NO_ALTERNATE_SCREEN;
   struct notcurses* nc;
   if((nc = notcurses_init(&opts, nullptr)) == nullptr){